refactor(oracle): remove invalid ballots
