+++ x/oracle/keeper/ballot.go
@@ -93,3 +93,42 @@ func (k Keeper) updateWhitelist(ctx sdk.Context, nextWhitelist []asset.Pair, cur
 		}
 	}
 }
+
+// RemoveInvalidBallots removes the ballots which have not reached the vote threshold
+// or which are not part of the whitelisted pairs anymore: example when params change during a vote period
+// but some votes were already made.
+//
+// ALERT: This function mutates pairBallotMap slice, it removes the ballot for the pair which is not passing the threshold
+// or which is not whitelisted anymore.
+func (k Keeper) RemoveInvalidBallots(
+	ctx sdk.Context,
+	pairBallotsMap map[asset.Pair]types.ExchangeRateBallots,
+) (map[asset.Pair]types.ExchangeRateBallots, map[asset.Pair]struct{}) {
+	whitelistedPairs := k.GetWhitelistedPairs(ctx)
+
+	whitelistedPairsMap := make(map[asset.Pair]struct{}, len(whitelistedPairs))
+	for _, pair := range whitelistedPairs {
+		whitelistedPairsMap[pair] = struct{}{}
+	}
+
+	totalBondedPower := sdk.TokensToConsensusPower(k.StakingKeeper.TotalBondedTokens(ctx), k.StakingKeeper.PowerReduction(ctx))
+	thresholdPower := k.VoteThreshold(ctx).MulInt64(totalBondedPower).RoundInt()
+
+	for pair, ballots := range pairBallotsMap {
+		// Ignore not whitelisted pairs
+		if _, exists := whitelistedPairsMap[pair]; !exists {
+			delete(pairBallotsMap, pair)
+			continue
+		}
+
+		// If the ballot is not passed, remove it from the whitelistedPairs set
+		// to prevent slashing validators who did valid vote.
+		if !isPassingVoteThreshold(ballots, thresholdPower) {
+			delete(whitelistedPairsMap, pair)
+			delete(pairBallotsMap, pair)
+			continue
+		}
+	}
+
+	return pairBallotsMap, whitelistedPairsMap
+}
