+++ x/oracle/keeper/ballot_test.go
@@ -4,15 +4,14 @@ import (
 	"sort"
 	"testing"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	"github.com/stretchr/testify/require"
-
 	"github.com/NibiruChain/collections"
-
 	"github.com/NibiruChain/nibiru/x/common/asset"
 	"github.com/NibiruChain/nibiru/x/common/denoms"
 	"github.com/NibiruChain/nibiru/x/oracle/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/staking"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestOrganizeAggregate(t *testing.T) {
@@ -188,4 +187,83 @@ func TestApplyWhitelist(t *testing.T) {
 		return nextWhitelist[i] < nextWhitelist[j]
 	})
 	require.Equal(t, nextWhitelist, input.OracleKeeper.WhitelistedPairs.Iterate(input.Ctx, collections.Range[asset.Pair]{}).Keys())
+}
+
+type VoteMap = map[asset.Pair]types.ExchangeRateBallots
+
+func TestRemoveInvalidBallots(t *testing.T) {
+	tests := []struct {
+		name    string
+		voteMap VoteMap
+	}{
+		{
+			name: "empty key, empty ballot",
+			voteMap: VoteMap{
+				"": types.ExchangeRateBallots{},
+			},
+		},
+		{
+			name: "nonempty key, empty ballot",
+			voteMap: VoteMap{
+				"xxx": types.ExchangeRateBallots{},
+			},
+		},
+		{
+			name: "nonempty keys, empty ballot",
+			voteMap: VoteMap{
+				"xxx":    types.ExchangeRateBallots{},
+				"abc123": types.ExchangeRateBallots{},
+			},
+		},
+		{
+			name: "mixed empty keys, empty ballot",
+			voteMap: VoteMap{
+				"xxx":    types.ExchangeRateBallots{},
+				"":       types.ExchangeRateBallots{},
+				"abc123": types.ExchangeRateBallots{},
+				"0x":     types.ExchangeRateBallots{},
+			},
+		},
+		{
+			name: "empty key, nonempty ballot, not whitelisted",
+			voteMap: VoteMap{
+				"": types.ExchangeRateBallots{
+					{Pair: "", ExchangeRate: sdk.ZeroDec(), Voter: sdk.ValAddress{}, Power: 0},
+				},
+			},
+		},
+		{
+			name: "nonempty key, nonempty ballot, whitelisted",
+			voteMap: VoteMap{
+				"x": types.ExchangeRateBallots{
+					{Pair: "x", ExchangeRate: sdk.Dec{}, Voter: sdk.ValAddress{123}, Power: 5},
+				},
+				asset.Registry.Pair(denoms.BTC, denoms.NUSD): types.ExchangeRateBallots{
+					{Pair: asset.Registry.Pair(denoms.BTC, denoms.NUSD), ExchangeRate: sdk.Dec{}, Voter: sdk.ValAddress{123}, Power: 5},
+				},
+				asset.Registry.Pair(denoms.ETH, denoms.NUSD): types.ExchangeRateBallots{
+					{Pair: asset.Registry.Pair(denoms.BTC, denoms.NUSD), ExchangeRate: sdk.Dec{}, Voter: sdk.ValAddress{123}, Power: 5},
+				},
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+
+		t.Run(tc.name, func(t *testing.T) {
+			testSetup, _ := setup(t)
+			ctx := testSetup.Ctx
+			oracleKeeper := testSetup.OracleKeeper
+
+			switch {
+			// case tc.err:
+			// TODO Include the error case when collections no longer panics
+			default:
+				assert.NotPanics(t, func() {
+					_, _ = oracleKeeper.RemoveInvalidBallots(ctx, tc.voteMap)
+				}, "voteMap: %v", tc.voteMap)
+			}
+		})
+	}
 }
